requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0
openai>=1.0.0
chromadb>=0.5
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
